@@ -9,22 +9,13 @@
     "validate": "office-toolbox validate -m manifest.xml"
   },
   "dependencies": {
-    "axios": "^0.21.1",
-    "core-js": "^3.6.5",
-    "element-ui": "^2.15.3",
-    "office-toolbox": "^0.1.1",
+    "axios": "^1.4.0",
+    "core-js": "^3.31.1",
+    "element-plus": "^2.3.7",
+    "office-toolbox": "^0.3.0",
     "speechmarkdown-js": "https://github.com/npocccties/speechmarkdown-js#ties-with-dist",
-    "vue": "^2.6.11",
-    "vuex": "^3.6.0"
-  },
-  "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^7.9.0",
-    "vue-template-compiler": "^2.6.11"
+    "vue": "^3.3.4",
+    "vuex": "^4.1.0"
   },
   "eslintConfig": {
     "root": true,
@@ -44,5 +35,9 @@
     "> 1%",
     "last 2 versions",
     "not dead"
-  ]
+  ],
+  "devDependencies": {
+    "@vue/cli-service": "^5.0.8",
+    "@vue/compiler-sfc": "^3.3.4"
+  }
 }
