@@ -1,32 +1,17 @@
-const execa = require('execa');
-const fs = require('fs');
-const path = require('path');
-const sm = require('speechmarkdown-js');
-const AWSPolly = require('aws-sdk/clients/polly');
-const mm = require('music-metadata');
-const axios = require('axios');
+import {execaCommand} from 'execa';
+import fs from 'fs';
+import path from 'path';
+import {PollyClient, SynthesizeSpeechCommand} from '@aws-sdk/client-polly';
+import mm from 'music-metadata-browser';
+import axios from 'axios';
 
-const {config} = require('./config.js');
-const {log4js, logger} = require('./log.js');
-const parse = require('./parse.js');
-const convert = require('./convert.js');
+import {config} from './config.js';
+import {log4js, logger} from './log.js';
+import {parse} from './parse.js';
+import {convert} from './convert.js';
+import {updateParams} from './audio.js';
 
-const {libDir, tikaJar, outputDir, tempDir} = config;
-
-async function getTika(filename) {
-  const filepath = path.parse(filename);
-  const cmd = `java -jar ${libDir}/${tikaJar} -x ${filename}`;
-
-  let xml;
-  if (filepath.ext === '.xml') {
-    xml = fs.readFileSync(filename, 'utf8');
-  } else {
-    logger.info(cmd);
-    const result = await execa.command(cmd);
-    xml = result.stdout;
-  }
-  return xml;
-}
+const {libDir, outputDir, tempDir} = config;
 
 function listSlideImages(path) {
   const slides = {};
@@ -74,107 +59,6 @@ function prepare(filepath, sections, vsuffix, asuffix) {
   });
 }
 
-function updateParam(prev, slide, req) {
-  // delay
-  if (typeof slide.delay === 'undefined') {
-    slide.delay = prev.delay;
-  }
-  // pad
-  if (typeof slide.pad === 'undefined') {
-    slide.pad = prev.pad;
-  }
-  // fade
-  if (typeof slide.fade === 'undefined') {
-    slide.fade = prev.fade;
-  }
-  // voice
-  if (typeof slide.voice !== 'undefined') {
-    req.VoiceId = slide.voice.trim();
-  }
-  // Takumi default engine
-  if (req.VoiceId == 'Takumi') {
-    req.Engine = 'neural';
-  } else {
-    delete req.Engine;
-  }
-  // engine
-  if (typeof slide.engine !== 'undefined') {
-    req.Engine = slide.engine.trim();
-  }
-  // sampleRate
-  if (typeof slide.sampleRate !== 'undefined') {
-    const trimmed = slide.sampleRate.trim();
-    if (polly_valid_samplerates.includes(trimmed)) {
-      req.SampleRate = trimmed;
-    }
-  }
-}
-
-function breakTime(sec) {
-  if (typeof sec === 'undefined' || sec <= 0) {
-    return '';
-  }
-  let ms = sec * 1000;
-  if (ms > 10000) {
-    ms = 10000;
-  }
-  return `<break time="${ms.toFixed()}ms"/>`;
-}
-
-function slideText(slide) {
-  const lines = slide.blocks.filter(b => b.type === 'text').map(b => b.content).flat(1);
-  const text = lines.join('\n').replace(/<\/?speak>/g,'');
-  const delay = breakTime(slide.delay);
-  const pad = breakTime(slide.pad);
-  return [delay, text, pad].join('');
-}
-
-const polly_valid_samplerates = ['8000', '16000', '22050', '24000'];
-
-const polly_defaults = {
-  OutputFormat: 'mp3',
-  SampleRate: config.sampleRate,
-  TextType: 'ssml',
-  VoiceId: config.voice,
-};
-
-async function createAudioFiles(slides, speech) {
-  const polly = new AWSPolly();
-
-  const req = {
-    ...polly_defaults
-  };
-  if (!polly_valid_samplerates.includes(req.SampleRate)) {
-    req.SampleRate = '22050';
-  }
-
-  let prev = {
-    delay: config.delay,
-    pad: config.pad,
-    fade: config.fade,
-  };
-
-  for(slide of slides) {
-    // update request
-    updateParam(prev, slide, req);
-    req.Text = speech.toSSML(slideText(slide));
-
-    // call amazon polly
-    let stream;
-    if (config.pollyProxy) {
-      const res = await axios.post(config.pollyProxy, req, {responseType: 'arraybuffer'});
-      stream = res.data;
-    } else {
-      const data = await polly.synthesizeSpeech(req).promise();
-      stream = data.AudioStream;
-    }
-    slide.duration = (await mm.parseBuffer(stream)).format.duration;
-    fs.writeFileSync(slide.audioFilename, stream);
-
-    prev = slide;
-  }
-}
-
 function ffmegLocation() {
   if (config.vcodec === 'libopenh264') {
     let ffmpeg = libDir + '/' + config.ffmpegCmd;
@@ -196,7 +80,7 @@ function ffmegLocation() {
 
 const vfCommon = ['fps=25','format=yuv420p'];
 const vcodec = `-c:v ${config.vcodec} ${config.voption}`;
-const acodec = '-c:a copy';
+const acodec = `-c:a ${config.acodec} ${config.aoption}`;
 const ffmpeg = ffmegLocation();
 
 function getVfOption(duration, option = {}) {
@@ -215,14 +99,14 @@ async function createVideo(slide, filename, option = {}) {
   const vfOption = getVfOption(slide.duration, option);
   const cmd = `${ffmpeg} -y -loop 1 -t ${slide.duration} -i ${slide.imageFilename} -i ${slide.audioFilename} ${vcodec} ${vfOption} ${acodec} ${filename}`;
   logger.info(cmd);
-  return execa.command(cmd);
+  return execaCommand(cmd);
 }
 
 async function concatenateVideo(topic, option = {}) {
   const vfOption = getVfOption(topic.duration, option);
-  const cmd = `${ffmpeg} -y -f concat -i ${topic.listFilename} ${vcodec} ${vfOption} ${acodec} ${topic.outputFilename}`;
+  const cmd = `${ffmpeg} -y -f concat -i ${topic.listFilename} ${vcodec} ${vfOption} -c:a copy ${topic.outputFilename}`;
   logger.info(cmd);
-  return execa.command(cmd);
+  return execaCommand(cmd);
 }
 
 function videoOption(slide) {
@@ -234,12 +118,12 @@ function videoOption(slide) {
 }
 
 async function createVideoFiles(sections) {
-  for (section of sections) {
-    for (topic of section.topics) {
+  for (const section of sections) {
+    for (const topic of section.topics) {
       const option = videoOption(topic.slides[0]);
       if (topic.slides.length > 1) {
         let duration = 0.0;
-        for (slide of topic.slides) {
+        for (const slide of topic.slides) {
           duration = duration + slide.duration;
           const result = await createVideo(slide, slide.videoFilename);
         }
@@ -295,35 +179,60 @@ function fatalError(e, message, sections = null) {
   exitProcess(-1);
 }
 
-async function ppt2video(filename) {
+async function createAudioFiles(slides) {
+  updateParams(slides);
+  const polly = new PollyClient();
+
+  for(const slide of slides) {
+    const {req} = slide;
+
+    // call amazon polly
+    let stream;
+    if (config.pollyProxy) {
+      const res = await axios.post(config.pollyProxy, req, {responseType: 'arraybuffer'});
+      stream = res.data;
+    } else {
+      const command = new SynthesizeSpeechCommand(req);
+      const res = await polly.send(command);
+      const chunks = []
+      for await (let chunk of res.AudioStream) {
+        chunks.push(chunk)
+      }
+      stream = Buffer.concat(chunks);
+    }
+    slide.duration = (await mm.parseBuffer(stream)).format.duration;
+    fs.writeFileSync(slide.audioFilename, stream);
+  }
+}
+
+export async function ppt2video(filename, getPptx) {
   const filepath = path.parse(filename);
-  const speech = new sm.SpeechMarkdown({platform: 'amazon-alexa'});
-  let xml, data;
+  let pptx, data;
 
   if (!fs.existsSync(filename)) {
     fatalError(null, 'can\'t find ' + filename);
   }
   logger.info('processing ' + filename);
 
-  // getting tika xml output
-  logger.trace('call getTika');
+  // getting pptx file
+  logger.trace('call getPptx');
   try {
-    xml = await getTika(filename);
+    pptx = await getPptx(filename);
   } catch(e) {
-    fatalError(e, 'can\'t get tika xml output.');
+    fatalError(e, 'can\'t get pptx file.');
     return;
   }
-  logger.trace(xml);
+  logger.trace(pptx.xml);
 
   // parse
   logger.trace('call parse');
   try {
-    data = await parse.parse(xml);
+    data = await parse(pptx);
   } catch(e) {
     fatalError(e, 'parse error');
     return;
   }
-  const {tika, slides, sections} = data;
+  const {slides, sections} = data;
   logger.trace(JSON.stringify(sections,null,2));
 
   // prepare output filename
@@ -334,37 +243,34 @@ async function ppt2video(filename) {
     return;
   }
 
+  if (!config.novideo) {
+    // create audio files
+    logger.trace('call createAudioFiles');
+    try {
+      await createAudioFiles(slides);
+    } catch(e) {
+      fatalError(e, 'failed to create audio files', sections);
+      return;
+    }
+
+    // create video files
+    logger.trace('call createVideoFiles');
+    try {
+      await createVideoFiles(sections);
+    } catch(e) {
+      fatalError(e, 'failed to create video files', sections);
+      return;
+    }
+  }
+
   // convert to import-json
   logger.trace('call convert');
   try {
-    const ij = convert.convert(tika, slides, sections);
+    const ij = convert(pptx, slides, sections);
     fs.writeFileSync(path.join(outputDir, filepath.name + '.json'), JSON.stringify(ij,null,2));
+    logger.info('success to craete import json file');
   } catch(e) {
     fatalError(e, 'failed to create import json file', sections);
-    return;
-  }
-
-  if (config.novideo) {
-    logger.info('success to craete import json file');
-    removeTempFiles(sections);
-    return;
-  }
-
-  // create audio files
-  logger.trace('call createAudioFiles');
-  try {
-    await createAudioFiles(slides, speech);
-  } catch(e) {
-    fatalError(e, 'failed to create audio files', sections);
-    return;
-  }
-
-  // create video files
-  logger.trace('call createVideoFiles');
-  try {
-    await createVideoFiles(sections);
-  } catch(e) {
-    fatalError(e, 'failed to create video files', sections);
     return;
   }
 
@@ -372,8 +278,4 @@ async function ppt2video(filename) {
   removeTempFiles(sections);
 
   exitProcess(0);
-}
-
-module.exports = {
-  ppt2video
 }
